requirements.txt
# Core Discord functionality
discord.py>=2.5.2

# Music player dependencies
yt-dlp>=2025.3.31
async-timeout>=5.0.1
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0

# File handling and HTTP
aiofiles>=24.1.0
aiohttp>=3.11.18

# Database dependencies
motor>=3.7.0
redis>=5.2.1
pymongo>=4.12.0

# Statistics dependencies
matplotlib>=3.7.1
pillow>=10.0.0